# Frontend/sssklv2-openapi.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "SSSKLv2 | v1",
        "contact": {
            "name": "Scouting Wilo",
            "email": "webmaster@scoutingwilo.nl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:7193/"
        }
    ],
    "paths": {
...
        "/v1/Leaderboard/{productId}": {
            "get": {
                "tags": [
                    "Leaderboard"
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LeaderboardEntryDto"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LeaderboardEntryDto"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LeaderboardEntryDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/Leaderboard/monthly/{productId}": {
            "get": {
                "tags": [
                    "Leaderboard"
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LeaderboardEntryDto"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LeaderboardEntryDto"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LeaderboardEntryDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/Leaderboard/12hour/{productId}": {
            "get": {
                "tags": [
                    "Leaderboard"
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LeaderboardEntryDto"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LeaderboardEntryDto"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LeaderboardEntryDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/TopUp": {
            "get": {
                "tags": [
                    "TopUp"
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TopUpDto"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TopUpDto"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TopUpDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TopUp"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/TopUpCreateDto"
                                    }
                                ]
                            }
                        },
                        "text/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/TopUpCreateDto"
                                    }
                                ]
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/TopUpCreateDto"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/TopUpDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TopUpDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TopUpDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/TopUp/personal": {
            "get": {
                "tags": [
                    "TopUp"
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TopUpDto"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TopUpDto"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TopUpDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/TopUp/{id}": {
            "get": {
                "tags": [
                    "TopUp"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/TopUpDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TopUpDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TopUpDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "TopUp"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "LeaderboardEntryDto": {
                "type": "object",
                "properties": {
                    "position": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "productName": {
                        "type": "string"
                    },
                    "amount": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    }
                }
            },
...
            "TopUpCreateDto": {
                "required": [
                    "userName"
                ],
                "type": "object",
                "properties": {
                    "userName": {
                        "type": "string"
                    },
                    "saldo": {
                        "minimum": 0.01,
                        "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
                        "type": [
                            "number",
                            "string"
                        ],
                        "format": "double"
                    }
                }
            },
            "TopUpDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "userName": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "saldo": {
                        "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
                        "type": [
                            "number",
                            "string"
                        ],
                        "format": "double"
                    },
                    "createdOn": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    },
...
}